6-size.c

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	char charType;
 
 	/*size of evaluates the size of a variable*/
-	printf("Size of int: %zu bytes\n", sizeof(intType));
-	printf("Size of float: %zu bytes\n", sizeof(floatType));
-	printf("Size of double: %zu bytes\n", sizeof(doubleType));
-	printf("Size of char: %zu byte\n", sizeof(charType));
+	printf("Size of int: %zu bytes\n",(unsigned long)sizeof(intType));
+	printf("Size of float: %zu bytes\n",(unsigned long)sizeof(floatType));
+	printf("Size of double: %zu bytes\n",(unsigned long)sizeof(doubleType));
+	printf("Size of char: %zu byte\n",(unsigned long)sizeof(charType));
 
 	return (0);
 }
